% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{polarity_index}
\alias{polarity_index}
\title{Polarity index calculation}
\usage{
polarity_index(organic_percent = numeric(), organic_modifier = character())
}
\arguments{
\item{organic_percent}{The organic modifier percentage at a time of elution or the isocratic percentage value}

\item{organic_modifier}{The organic modifier for which polarity index is calculated (either "MeOH" or "MeCN")}
}
\value{
Polarity index (units?)
}
\description{
This function calculates the polarity index based on organic modifier and organic modifier percentage
}
\examples{
polarity_index_suspect <- polarity_index(organic_percent = 80, 
                                         organic_modifier = "MeOH");
}
