% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifi_utils.R
\name{read_mzML}
\alias{read_mzML}
\title{Convert mzml to ms files (Unifi Export)}
\usage{
read_mzML(path_to_folder, create_ms = TRUE)
}
\arguments{
\item{path_to_folder}{The path to the folder containing the .mzml files exported from Unifi}

\item{create_ms}{logical: choose to create .ms files (default = TRUE)}
}
\value{
Spectra object (MSnbase)
}
\description{
Use this function to convert processed mzml spectra files, exported from Unifi, to a spectra object (MSnbase), and generate .ms files for use in SIRIUS, MetFrag, and MassBank.
}
