% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMILES2descriptors_fingerprints.R
\name{SMILES2SIRIUS_fp}
\alias{SMILES2SIRIUS_fp}
\title{Calculate the true (binary) SIRIUS fingerprint for given SMILES notations}
\usage{
SMILES2SIRIUS_fp(SMILES_list, write_raw_fp = T, mode = "raw")
}
\arguments{
\item{SMILES_list}{data.frame: Data frame containing SMILES notations}

\item{write_raw_fp}{logical: Choose whether to write out the file (.tsv) that contains raw
(all possible features, 8924) SIRIUS fingerprints (default = TRUE)}

\item{mode}{character: Specifies which SIRIUS fingerprint features will be returned:
- 'pos': Positive ionization mode features (3878)
- 'neg': Negative ionization mode features (4072)
- 'overlapping': Features common to both ionization modes (3494)
- 'raw': Raw SIRIUS fingerprint (all 8925 features)}
}
\value{
data.frame: Contains SMILES and calculated fingerprint features based on the selected mode
}
\description{
This function reads structural patterns from files downloaded from https://github.com/boecker-lab/sirius,
which correspond to SIRIUS:CSI:FingerID outputted fingerprint features.
After that it generates binary SIRIUS fingerprints for all input SMILES notations.
}
\author{
Ida Rahu
}
